refactor VotingProps
