+++ packages/lesswrong/components/votes/votingProps.ts
@@ -1,0 +1,8 @@
+
+export interface VotingProps<T extends VoteableTypeClient> {
+  vote: (props: { document: T; voteType: string | null; extendedVote?: any; currentUser: UsersCurrent; }) => void;
+  collectionName: VoteableCollectionName;
+  document: T;
+  baseScore: number;
+  voteCount: number;
+}
